separation of the code that handles cell overlap values on the pane
